fix: separated route functionalities, added wiki router
