+++ server/api/games.js
@@ -1,5 +1,5 @@
 const router = require('express').Router()
-const { getDate, getRandomNums, underTitleize, titleize } = require('./utils')
+const { underTitleize } = require('./utils')
 const { db } = require('../db/config')
 const axios = require('axios')
 module.exports = router
@@ -23,18 +23,10 @@ router.get('/', async (req, res, next) => {
   }
 })
 
-//creates a new game instance, called by generate game
+//creates a new game instance in db, called by generate game
 router.post('/', async (req, res, next) => {
   try {
-    // Get top articles of the day from WikiPedia
-    const date = getDate()
-    const topArticles = await axios.get(`https://wikimedia.org/api/rest_v1/metrics/pageviews/top/en.wikipedia.org/all-access/${date[0]}/${date[1]}/${date[2]}`)
-    const articles = topArticles.data.items[0].articles
-    // Assign random top article to start and target articles of the game
-    const randomNums = getRandomNums()
-    const start = titleize(articles[randomNums[0]].article)
-    const target = titleize(articles[randomNums[1]].article)
-    //Create a new game instance in Firebase
+    const { start, target } = req.body
     const newGameId = await db.ref('Games/').push().key
     await db.ref('Games/' + newGameId).set({
       gameId: newGameId,
@@ -44,7 +36,7 @@ router.post('/', async (req, res, next) => {
       players: true,
       clickInfo: true
     })
-    res.send({ newGameId, start, target })
+    res.send({ newGameId })
   } catch (err) {
     next(err)
   }
@@ -61,17 +53,18 @@ router.put('/', async (req, res, next) => {
   }
 })
 
-// fetches current game start and target and html, called by join game
+// fetches current game start and target, called by join game
 router.get('/:gameId', (req, res, next) => {
-  const gameId = req.params.gameId
-  const gameRef = db.ref(`Games/${gameId}`)
-  gameRef.on('value', async (snapshot) => {
-    const data = snapshot.val()
-    const title = underTitleize(data.start)
-    const response = await axios.get(`https://en.wikipedia.org/api/rest_v1/page/html/${title}`)
-    console.log('HTML', response.data)
-    res.send({ start: data.start, target: data.target, html: response.data })
-  })
+  try {
+    const gameId = req.params.gameId
+    const gameRef = db.ref(`Games/${gameId}`)
+    gameRef.on('value', async (snapshot) => {
+      const data = snapshot.val()
+      res.send({ start: data.start, target: data.target })
+    })
+  } catch (err) {
+    next(err)
+  }
 })
 
 // creates a new player in the current game with player game info and adds the game to user's history, called from join a game
